Pressure position.
Various refactoring.  And blink when ext-osc is on.

+++ src/clocky32.rs
@@ -22,6 +22,6 @@ mod tsc;
 mod utils;
 mod vcell;
 
-const LCD_WIDTH: usize = 4;
+const LCD_WIDTH: u32 = 4;
 
 use clocky::*;
